Implement simple country identification based on countries mentioned in article content
